requirements2.txt
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-crispy-forms==1.9.1
django-formtools==2.2
django-import-export==2.2.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.7
django-simple-captcha==0.5.12
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.1
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.2
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
view==0.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
mysqlclient==2.0.3